added all files for a linked list in c++
attached is code for a linked list in c++. it is intended to be a double linked list, and it will function as such until the merge sort algorithm (included) is applied. for some reason it loses track of one of the previous pointers and therefore breaks the double link, however the forward links function perfectly.